+++ examples/monkey_saddle/monkey_saddle_constraints.py
@@ -18,6 +18,8 @@ from jax_fdm.equilibrium import constrained_fdm
 from jax_fdm.optimization import TrustRegionConstrained
 from jax_fdm.optimization import OptimizationRecorder
 
+from jax_fdm.parameters import EdgeForceDensityParameter
+
 from jax_fdm.goals import NodeResidualForceGoal
 
 from jax_fdm.constraints import EdgeLengthConstraint
@@ -49,7 +51,7 @@ length_max = 4.0  # maximum allowed edge length for length constraint
 
 optimizer = TrustRegionConstrained  # optimization algorithm
 maxiter = 500  # optimizer maximum iterations
-tol = 1e-3  # optimizer tolerance
+tol = 1e-2  # optimizer tolerance
 
 record = True  # True to record optimization history of force densities
 export = False  # export result to JSON
@@ -77,7 +79,7 @@ mesh.collect_polyedges()
 print("Densified mesh:", mesh)
 
 # ==========================================================================
-# Define support conditions
+# Define anchor conditions
 # ==========================================================================
 
 polyedge2length = {}
@@ -86,15 +88,15 @@ for pkey, polyedge in mesh.polyedges(data=True):
         length = sum([mesh.edge_length(u, v) for u, v in pairwise(polyedge)])
         polyedge2length[tuple(polyedge)] = length
 
-supports = []
+anchors = []
 n = sum(polyedge2length.values()) / len(polyedge2length)
 for polyedge, length in polyedge2length.items():
     if length < n:
-        supports += polyedge
+        anchors += polyedge
 
-supports = set(supports)
+anchors = set(anchors)
 
-print("Number of supported nodes:", len(supports))
+print("Number of anchored nodes:", len(anchors))
 
 # ==========================================================================
 # Compute assembly sequence (simplified)
@@ -105,7 +107,7 @@ corners = set([vkey for vkey in mesh.vertices() if mesh.vertex_degree(vkey) == 2
 adjacency = mesh.adjacency
 weight = {(u, v): 1.0 for u in adjacency for v in adjacency[u]}
 
-for vkey in supports:
+for vkey in anchors:
     if vkey in corners:
         steps[vkey] = 0
     else:
@@ -122,13 +124,13 @@ steps = {vkey: max_step - step for vkey, step in steps.items()}
 # ==========================================================================
 
 nodes = [mesh.vertex_coordinates(vkey) for vkey in mesh.vertices()]
-edges = [(u, v) for u, v in mesh.edges() if u not in supports or v not in supports]
+edges = [(u, v) for u, v in mesh.edges() if u not in anchors or v not in anchors]
 network0 = FDNetwork.from_nodes_and_edges(nodes, edges)
 
 print("FD network:", network0)
 
 # data
-network0.nodes_supports(supports)
+network0.nodes_anchors(anchors)
 network0.nodes_loads([px, py, pz], keys=network0.nodes())
 network0.edges_forcedensities(q=q0)
 
@@ -154,12 +156,21 @@ network0 = fdm(network0)
 network0.print_stats()
 
 # ==========================================================================
+# Define parameters
+# ==========================================================================
+
+parameters = []
+for edge in network0.edges():
+    parameter = EdgeForceDensityParameter(edge, qmin, qmax)
+    parameters.append(parameter)
+
+# ==========================================================================
 # Define goals
 # ==========================================================================
 
 # reaction forces
 goals = []
-for key in network0.nodes_supports():
+for key in network0.nodes_anchors():
     step = steps[key]
     reaction = (1 - step / max_step) ** r_exp * (rmax - rmin) + rmin
     goal = NodeResidualForceGoal(key, reaction)
@@ -171,7 +182,6 @@ for key in network0.nodes_supports():
 
 constraints = None
 if add_constraints:
-
     constraints = []
     for edge in network0.edges():
         constraint = EdgeLengthConstraint(edge, bound_low=length_min, bound_up=length_max)
@@ -188,13 +198,14 @@ loss = Loss(squared_error)
 # Solve constrained form-finding problem
 # ==========================================================================
 
-recorder = None
-if record:
-    recorder = OptimizationRecorder()
+optimizer = optimizer()
+
+recorder = OptimizationRecorder(optimizer) if record else None
 
 network = constrained_fdm(network0,
-                          optimizer=optimizer(),
+                          optimizer=optimizer,
                           loss=loss,
+                          parameters=parameters,
                           constraints=constraints,
                           maxiter=maxiter,
                           tol=tol,