+++ src/jax_fdm/goals/edge/angle.py
@@ -1,10 +1,11 @@
 import numpy as np
+
 import jax.numpy as jnp
 
 from jax_fdm.geometry import angle_vectors
 
 from jax_fdm.goals import ScalarGoal
-from jax_fdm.goals.edgegoal import EdgeGoal
+from jax_fdm.goals.edge import EdgeGoal
 
 
 class EdgeAngleGoal(ScalarGoal, EdgeGoal):