+++ src/jax_fdm/goals/network/loadpath.py
@@ -1,7 +1,8 @@
 import jax.numpy as jnp
 
 from jax_fdm.goals import ScalarGoal
-from jax_fdm.goals.networkgoal import NetworkGoal
+
+from jax_fdm.goals.network import NetworkGoal
 
 
 class NetworkLoadPathGoal(ScalarGoal, NetworkGoal):
@@ -16,6 +17,6 @@ class NetworkLoadPathGoal(ScalarGoal, NetworkGoal):
         """
         The current load path of the network.
         """
-        load_path = jnp.sum(jnp.abs(eq_state.lengths * eq_state.forces))
+        load_path = jnp.sum(jnp.abs(jnp.multiply(eq_state.lengths, eq_state.forces)))
 
         return jnp.atleast_1d(load_path)