+++ src/jax_fdm/goals/node/__init__.py
@@ -1,0 +1,12 @@
+from .node import *  # noqa F403
+from .point import *  # noqa F403
+from .coordinates import *  # noqa F403
+from .normal import *  # noqa F403
+from .tangent import *  # noqa F403
+from .line import *  # noqa F403
+from .segment import *  # noqa F403
+from .plane import *  # noqa F403
+from .residual import *  # noqa F403
+
+
+__all__ = [name for name in dir() if not name.startswith('_')]